Move common implementations for postgres and SQLite to TableImplementation
